Move Address common components into a shared folder
